Refactored Loading and LinkTab into reusable folder
