--- src/components/NavBar/NavBarStyle.js
@@ -1,12 +1,0 @@
-import styled from "styled-components";
-import AddIcon from "@mui/icons-material/Add";
-
-export const NavBarAddIcon = styled(AddIcon)`
-padding: "0px",
-margin: "0px",
-minWidth: "0px",
-"&:hover": {
-  backgroundColor: "lightgrey",
-  borderRadius: "50%",
-}
-`;
