+++ src/views/Antimedium/Antimedium.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import Loading from "../../components/Loading/Loading";
+import Loading from "../../components/reusable-components/Loading/Loading";
 
 const Antimedium = () => {
   return (
